Feat: Add AdjusterMonkey.Core.js
Add «Core/AdjusterMonkey.Core.js» to the project as a Tampermonkey script that will serve as the core from which all AdjusterMonkey scripts are built to enhance specific websites or web apps. Begin writing an AdjusterMonkeyGui class, which will be used to provide a GUI to the user for interacting with a specific AdjusterMonkey script, by implementing its constructor.