Feat: AdjusterMonkeyGui → Gui creation process
Write additional constructor methods to the «AdjusterMonkeyGui» class in «Core/AdjusterMonkey.Core.js» to create the GUI within the DOM and append it to the children of the target node. Create the GUI with the following components: logo, tagline, buttons for changing the position to which the GUI snaps, a button to expand and collapse the GUI, and an area for placing commands implemented as buttons. As part of this process, write additional default settings for parameters that are important when creating the GUI. Plan for two new core commands, one to provide information about the AdjusterMonkey Core script, the other to quickly obtain a reference link to the web page currently being browsed.